% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_ids_from_area.R
\name{gmba_ids_from_area}
\alias{gmba_ids_from_area}
\title{Get mountain range IDs from area}
\usage{
gmba_ids_from_area(area, calc)
}
\arguments{
\item{area}{Numeric, threshold value of area in size, given in km^2}

\item{calc}{Character string, should the area be taken as \emph{"max"}imum or \emph{"min"}imum?}
}
\value{
Character vector of the GMBA Inventory v2.0 mountain range IDs having a smaller or bigger
size in area as submitted
}
\description{
Get GMBA Inventory v2.0 mountain range IDs of mountain ranges that have a certain size in area
}
\examples{
\dontrun{
rangeids <- gmba_ids_from_area(5000, "max")
rangeids <- gmba_ids_from_area(5000, "min")
}
}
