% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_ids_from_country.R
\name{gmba_ids_from_country}
\alias{gmba_ids_from_country}
\title{Get mountain range IDs from country codes}
\usage{
gmba_ids_from_country(iso)
}
\arguments{
\item{iso}{Character vector of one or more ISO-3 country code(s). If codes are not known, check
\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{Wikipedia} or use the R package
\href{https://cran.r-project.org/web/packages/countrycode/index.html}{countrycode} to convert
country names}
}
\value{
Character vector of the GMBA Inventory v2.0 mountain range IDs that are located within
the input country codes
}
\description{
Get GMBA Inventory v2.0 mountain range IDs that are within one or more ISO-3 country codes
}
\examples{
\dontrun{
rangeid <- gmba_ids_from_country("che")
rangeids <- gmba_ids_from_country(c("che", "deu"))
}
}
