% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_ids_from_elevation.R
\name{gmba_ids_from_elevation}
\alias{gmba_ids_from_elevation}
\title{Get mountain range IDs from elevation range}
\usage{
gmba_ids_from_elevation(elevation_low, elevation_high)
}
\arguments{
\item{elevation_low}{Numeric, lower limit of an elevation range}

\item{elevation_high}{Numeric, higher limit of an elevation range}
}
\value{
Character vector of the GMBA Inventory v2.0 mountain range IDs associated
covering the input elevation range
}
\description{
Get GMBA Inventory v2.0 mountain range IDs that are within a certain mountain elevation range
}
\examples{
\dontrun{
rangeids <- gmba_ids_from_elevation(4000, 5000)
}
}
