% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_ids_from_names.R
\name{gmba_ids_from_names}
\alias{gmba_ids_from_names}
\title{Get mountain range IDs from mountain range names}
\usage{
gmba_ids_from_names(rangename_vector)
}
\arguments{
\item{rangename_vector}{Character vector of GMBA Inventory v2.0 mountain range names.
Use \code{gmba_search_names()} if you do not know the inventory names needed}
}
\value{
Character vector of the GMBA Inventory v2.0 mountain range IDs associated
with the input names
}
\description{
Get GMBA Inventory v2.0 mountain range IDs from known mountain range names.
}
\examples{
\dontrun{
rangeid <- gmba_ids_from_names("Black Forest")
rangeids <- gmba_ids_from_names(c("Black Forest", "European Alps"))
}
}
