% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_ids_from_points.R
\name{gmba_ids_from_points}
\alias{gmba_ids_from_points}
\title{Get mountain range IDs from points coordinates}
\usage{
gmba_ids_from_points(xy_dataframe, range_selection = "all", manual = NULL)
}
\arguments{
\item{xy_dataframe}{A dataframe containing point coordinates,
with columns "x" and "y" i.e. longitude and latitude}

\item{range_selection}{Character string, the options are:
\itemize{
\item{\emph{"all"} = all level ranges, cross-scale (returns lists of ids!) (default)}
\item{\emph{"basic"} = basic mountain ranges only}
\item{\emph{"300"} = ranges included in the 300 pre-selection}
\item{\emph{"100"} = ranges included in the 100 pre-selection}
\item{\emph{"level1"}-\emph{"level10"} = ranges pertaining to the respective level
of the inventory hierarchy}
\item{\emph{"manual"} = range IDs provided with \code{manual}}}}

\item{manual}{Character string containing the IDs of manually selected mountain ranges.
Only needed when \code{range_selection} is set to \emph{manual}. Default is set to \emph{NULL}}
}
\value{
Character vector of the GMBA Inventory v2.0 mountain range IDs associated
with the input names
}
\description{
Get GMBA Inventory v2.0 mountain range IDs based on xy point coordinates.
}
\examples{
\dontrun{
rangeid_vector <- gmba_ids_from_points(xy, "300")
rangeid_vector <- gmba_ids_from_points(xy, "manual", c("11175", "10001"))
rangeid_vector <- gmba_ids_from_points(xy, "manual",
                                       gmba_ids_from_names(c("Black Forest", "European Alps")))
}
}
