% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_ids_from_selectiontool.R
\name{gmba_ids_from_selectiontool}
\alias{gmba_ids_from_selectiontool}
\title{Get mountain range IDs from the Selection Tool}
\usage{
gmba_ids_from_selectiontool(local, overlap = FALSE)
}
\arguments{
\item{local}{Character string containing the path to the Selection Tool on the local drive}

\item{overlap}{Logical, set if to output those selected mountain ranges from the Selection Tool
that overlap with other (only those!) or not. Default is set to FALSE}
}
\value{
Vector of the GMBA Inventory v2.0 mountain range names associated with the input IDs
}
\description{
Get your pre-selected GMBA Inventory v2.0 mountain range IDs from the \href{https://github.com/GMBA-biodiversity/Inventory/blob/main/Gmba_Inventory_v2.0_Selection_Tool_20200330.xlsx}{Selection Tool} Excel file.
}
\examples{
\dontrun{
selection <- gmba_ids_from_selectiontool("Selection_Tool.xlsx")
selection_overlapping <- gmba_ids_from_selectiontool("Selection_Tool.xlsx",
                                                     overlap = TRUE)
}
}
