% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_ipbes_maj_str.R
\docType{data}
\name{gmba_ipbes_maj_str}
\alias{gmba_ipbes_maj_str}
\title{IPBES (sub)region data for gmbaR}
\format{
Dataframe with three columns: GMBA_V2_ID, IPBES_maj, IPBES_str
}
\usage{
data(gmba_ipbes_maj_str)
}
\description{
Data for use in \emph{gmba_ids_from_spi( )}. Information on which IPBES subregion each
mountain range of the GMBA inventory v2.0 is located in.
}
\references{
\href{https://doi.org/10.5281/zenodo.3928281}{doi.org/10.5281/zenodo.3928281}
}
\keyword{datasets}
