% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_kbapa.R
\name{gmba_kbapa}
\alias{gmba_kbapa}
\title{Filter the GMBA KBA/PA master file}
\usage{
gmba_kbapa(fileobject, variable, option)
}
\arguments{
\item{fileobject}{Name of the master file object in the R studio environment, without quotes}

\item{variable}{Character string of variable field/column from the master file, all except "ResultValue"}

\item{option}{Character string of the option to filter}
}
\value{
Dataframe filtered from given master file object
}
\description{
Filter the results from the GMBA results on PA coverage of mountain KBAs. Currently for internal use only.
}
\examples{
\dontrun{
 waffledata <- gmba_kbapa("Calculation", "rounded")
}
}
