% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_lower_id_from_higher.R
\name{gmba_lower_id_from_higher}
\alias{gmba_lower_id_from_higher}
\title{Get a lower mountain range ID from a higher mountain range ID}
\usage{
gmba_lower_id_from_higher(rangeid, lowerlevel_numeric, method)
}
\arguments{
\item{rangeid}{String (character or numeric) of a single GMBA Inventory v2.0 mountain range ID}

\item{lowerlevel_numeric}{A single number (as numeric), indicating what lower level to be given}

\item{method}{Character string, the options are:
\itemize{
\item{\emph{"parent"} = a certain lower range of range ID}
\item{\emph{"steps"} = number of lower levels from range ID}}
See the \href{https://github.com/GMBA-biodiversiry/gmbaR/blob/master/README.md}{gmbaR README} for a visual explanation of the two methods}
}
\value{
Character string of the lower GMBA Inventory v2.0 mountain range ID associated
with the input ID
}
\description{
Get a certain lower GMBA Inventory v2.0 mountain range ID from a higher mountain range ID.\cr\cr
Within the GMBA Inventory v2.0 hierarchy, "lower" levels are spatially high,
such as on the continental level. Consequently, "higher" levels are those of spatial detail,
such as individual mountain ranges within a mountain system.
}
\examples{
\dontrun{
Europe <- gmba_lower_id_from_higher(11175, 2, method = "parent")
CentralUplands <- gmba_lower_id_from_higher(11175, 2, method = "steps")
}
}
