% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_names_from_ids.R
\name{gmba_names_from_ids}
\alias{gmba_names_from_ids}
\title{Get mountain range names from mountain range IDs}
\usage{
gmba_names_from_ids(rangeid_vector)
}
\arguments{
\item{rangeid_vector}{Vector (character or numeric) of GMBA Inventory v2.0 mountain range IDs}
}
\value{
Vector of the GMBA Inventory v2.0 mountain range names associated with the input IDs
}
\description{
Get GMBA Inventory v2.0 mountain range names from known mountain range IDs.
}
\examples{
\dontrun{
 rangenames <- gmba_names_from_ids(c(11175, 10001))
 rangenames <- gmba_names_from_ids(c("11175", "10001"))
}
}
