% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_read.R
\name{gmba_read}
\alias{gmba_read}
\title{Read the GMBA Inventory v2.0 to R}
\usage{
gmba_read(source = "web", local = NULL)
}
\arguments{
\item{source}{Character string, options are \emph{web} (Default) or \emph{local}}

\item{local}{Character string containing the path to the inventory shapefile on the local drive.
Only needed when \code{source} is set to \emph{local}. Default is set to \emph{NULL}}
}
\value{
\code{gmba_inv()}, a function containing the GMBA Inventory v2.0
}
\description{
Read the GMBA Inventory v2.0 to R, from the web or the local drive.
}
\examples{
\dontrun{
gmba_read(source = "local", local = "GMBA_Inventory_V2_All_Attributes.shp")
}
}
\references{
\url{https://www.earthenv.org/mountains}
}
