% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_search_names.R
\name{gmba_search_names}
\alias{gmba_search_names}
\title{Search for mountain range names}
\usage{
gmba_search_names(part_of_rangename)
}
\arguments{
\item{part_of_rangename}{Character string of part of a mountain range name}
}
\value{
Character vector of the GMBA Inventory v2.0 mountain range names
associated with the input name part
}
\description{
Search GMBA Inventory v2.0 mountain range names based on parts of names.
}
\examples{
\dontrun{
 gmba_search_names("forest")
}
}
