% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_select.R
\name{gmba_select}
\alias{gmba_select}
\title{Select mountain ranges from the GMBA Inventory v2.0}
\usage{
gmba_select()
}
\value{
Depending on the output format selection, see details
}
\description{
Interactively filter and select mountain ranges from the GMBA Inventory v2.0.
This function can be considered as the \href{url}{Selection Tool} implemented in R.
}
\details{
{First, you are given four options to filter the inventory. After filtering, you can select
all or individual ranges from the filter. A 'stop' as response to each question stops
the function from running.\cr\cr
The filter options are:
\enumerate{
\item{By mountain range name (English only, for searching mountain range names in other
languages please use \code{gmba_search_names( )})}
\item{By country, using ISO-3 codes}
\item{By IPBES region or subregion, using the \emph{"str"} method from \code{gmba_ids_from_spi( )}}
\item{By IPCC AR6 region, using the \emph{"str"} method from \code{gmba_ids_from_spi( )}}}
Then, the selection options are:
\enumerate{
\item{All of the filtered mountain ranges}
\item{By number(s), as displayed after the last filter}}
Finally, the output options are:
\enumerate{
\item{An \code{sf} object containing the selected mountain ranges}
\item{A character vector with the GMBA Inventory v2.0 database names of the selected mountain ranges}
\item{A character vector with the GMBA Inventory v2.0 map names of the selected mountain ranges}
\item{A character vector with the GMBA Inventory v2.0 IDs of the selected mountain ranges}}
}
}
\examples{
\dontrun{
 gmba_select()
}
}
