% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{gmbaR}
\alias{gmbaR-package}
\alias{gmbaR}
\title{gmbaR}
\description{
The GMBA Mountain Inventory v2.0 is an inventory of 8329 mountain ranges across the world,
stored in a shapefile with a growing attribute table. Here, we provide a few R functions
to explore the inventory and select mountain ranges based on user define criteria for
subsequent analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GMBA-biodiversity/gmbaR}
  \item Report bugs at \url{https://github.com/GMBA-biodiversity/gmbaR/issues}
}

}
\keyword{internal}
