% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_ipcc_ar6_maj_str.R
\docType{data}
\name{gmba_ipcc_ar6_maj_str}
\alias{gmba_ipcc_ar6_maj_str}
\title{IPCC AR6 region data for gmbaR}
\format{
Dataframe with three columns: GMBA_V2_ID, IPCC_maj, IPCC_str
}
\usage{
data(gmba_ipcc_ar6_maj_str)
}
\description{
Data for use in \emph{gmba_ids_from_spi( )}. Information on which IPCC AR6 region each
mountain range of the GMBA inventory v2.0 is located in.
}
\references{
\href{https://doi.org/10.5194/essd-12-2959-2020}{doi.org/10.5194/essd-12-2959-2020}
}
\keyword{datasets}
