% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_kbapa.R
\name{gmba_kbapa}
\alias{gmba_kbapa}
\title{Filter the GMBA KBA/PA master file}
\usage{
gmba_kbapa(fileobject, options)
}
\arguments{
\item{fileobject}{Name of the master file object in the R studio environment, without quotes}

\item{option}{List of vectors of character strings containing the options to filter, in the format \code{list(c("variable", "filterterm"), ...)}.
The filter vectors are applied in the order given. The variable \emph{"ResultValue"} can not be filtered.
Variables that can be filtered: "ID", "ID_underscore", "Year", "UnitOfAnalysis", "Landscape", "Country", "Mountain", "Name", "Metric", "Definition", "Calculation", "Unit", and "FilterString".
For variables \emph{"Name"} and \emph{"Metric"}, pattern matching is used, using \code{grepl( )}.}
}
\value{
Dataframe filtered from given master file object
}
\description{
Filter the results from the GMBA results on PA coverage of mountain KBAs. Currently for internal use only.
}
\examples{
\dontrun{
 waffledata <- gmba_kbapa(data, "Calculation", "rounded")
}
}
