% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_read.R
\name{gmba_read}
\alias{gmba_read}
\title{Read the GMBA Inventory v2.0 to R}
\usage{
gmba_read(source = "web", extend = "standard", local = NULL, timeout = 300)
}
\arguments{
\item{source}{Character string, options are \emph{"web"} (Default) or \emph{"local"}}

\item{extend}{Character string, options are \emph{"standard"} (Default) or \emph{"broad"}. Check \url{https://www.earthenv.org/mountains}
for the details on the options. Ignored when \code{source} is set to \emph{"local"}}

\item{local}{Character string containing the path to the inventory shapefile on the local drive.
Only needed when \code{source} is set to \emph{"local"}. Default is set to \emph{NULL}}

\item{timeout}{Numeric time in seconds, after what time period the function should stop trying to download the GMBA Inventory v2.0. Default is set to \emph{300} i.e. 5 minutes}
}
\value{
\code{gmba_inv()}, a function containing the GMBA Inventory v2.0
}
\description{
Read the GMBA Inventory v2.0 (standard extend, all layers) to R, from the web or the local drive. This function is to be used
without an assignment operator (e.g. "<-").
}
\examples{
\dontrun{
gmba_read()
gmba_read("web", extend = "broad")
gmba_read(source = "local", local = "GMBA_Inventory_v2.0.shp")
}
}
\references{
\url{https://www.earthenv.org/mountains}
}
