% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmba_select.R
\name{gmba_select}
\alias{gmba_select}
\title{Select mountain ranges from the GMBA Inventory v2.0}
\usage{
gmba_select()
}
\value{
Depending on the output format selection, see details
}
\description{
Interactively filter and select mountain ranges from the GMBA Inventory v2.0.
This function represents the \href{url}{Selection Tool} integrated in R.
}
\details{
{First, you are given four options to filter the inventory. After filtering, you can select
all or individual ranges from the filter. A 'stop' as response to each question stops
the function from running.\cr\cr
The filter options are:
\enumerate{
\item{By mountain range name}
\item{By country}
\item{By IPBES region or subregion, based on \href{https://doi.org/10.5281/zenodo.3928281}{doi:10.5281/zenodo.3928281}}
\item{By IPCC AR6 region, based on \href{https://doi.org/10.5194/essd-12-2959-2020}{doi:10.5194/essd-12-2959-2020}}}
Then, the selection options are:
\enumerate{
\item{All of the filtered mountain ranges}
\item{By number(s), as displayed after the last filter}}
Finally, the output options are:
\enumerate{
\item{An \code{sf} object containing the selected mountain ranges}
\item{A character vector with the names of the selected mountain ranges}
\item{A character vector with the GMBA v2.0 IDs of the selected mountain ranges}}
}
}
\examples{
\dontrun{
 gmba_select()
}
}
